Update bicubicInterpolationTest function implementation
Update bicubicInterpolationTest function implementation

+++ main.cpp
@@ -347,7 +347,7 @@ void bicubicInterpolationTest()
     image1.setAllValue(1);
     std::cout << "Width: " + std::to_string(image1.getWidth()) + "\n";
     std::cout << "Height: " + std::to_string(image1.getHeight()) + "\n";
-    image1.at(1, 1) = 100;
+    image1.at(static_cast<std::size_t>(1), static_cast<std::size_t>(1)) = 100;
     image1.print();
 
     auto image2 = TinyDIP::copyResizeBicubic(image1, 12, 12);
